// client/archival-material-client/src/models/TypeOfRecordEnum.ts
enum TypeOfRecordEnum{
    UNSET = "UNSET",
    MATRIKA = "MATRIKA",
    SCITACI_OPERATOR = "SCITACI_OPERATOR",
    INDIKACNA_SKICA = "INDIKACNA_SKICA",
    POZEMKOVA_KNIHA = "POZEMKOVA_KNIHA",
    LANOVY_REJSTRIK = "LANOVY_REJSTRIK",
    URBAR = "URBAR",
    VEDUT = "VEDUT",
    MAPA = "MAPA",
    REKTIFIKACNI_AKTA = "REKTIFIKACNI_AKTA",
    BERNI_RULA = "BERNI_RULA",
    SOUPIS_PODDANYCH_DLE_VIRY = "SOUPIS_PODDANYCH_DLE_VIRY",
    OBEC = "OBEC",
    VSE_BEZ_MATRIK = "VSE_BEZ_MATRIK"
}

export default TypeOfRecordEnum;